requirements_dev.txt
openslide
pillow
scikit-image
matplotlib
